requirements.txt
Pillow==3.4.1
selenium==2.53.6
